refactor(lib): graph access toolkit - one shared Surreal module + named command queries
Consolidate the five SurrealDB access primitive modules into one documented
module and move the last inline command SQL behind named query functions.
Behavior-preserving: CLI output bytes, --json envelopes, and generated SQL
semantics are unchanged.

Graph Access Toolkit:
- packages/lib/src/shared/surreal.ts is the single module of safe Storage
  Backend access primitives, with five clearly-sectioned parts moved verbatim
  from surql.ts (literals/escaping), row-fields.ts (typed row field access),
  record-select.ts (record selection + the 3.0.x id-IN quirk docs),
  statement-exec.ts (statement execution), and derive-keys.ts (record-id key
  derivation). Every old file is now a /** @deprecated import from
  ./surreal.ts */ re-export shim, so no existing import breaks. The five test
  files are merged into surreal.test.ts (91 tests, green). graph-query.ts /
  query.ts stay their own typed read DSL module.
- Broke the json.ts -> shared/surreal.ts -> db.ts -> config -> runtime-state
  -> json.ts module-eval cycle introduced by surreal.ts's new db.ts dependency:
  runtime-state.ts now inlines the byte-identical 2-space stringify instead of
  importing json.ts's prettyPrint (a real JSON-IO boundary).

Named command queries:
- improve verdict list/show SQL moved to improve/verdicts.ts (listVerdicts /
  showVerdict). These carry per-row correlated checkpoint subqueries and a
  nested object/array shape, so per the 2026-05-21 decision they stay raw-SQL
  named functions with typed row interfaces rather than being forced through
  defineQuery/runQuery.
- costs summary + pricing SQL moved to dashboard/cost-summary-query.ts
  (fetchCostSummaryRollup / fetchPricingRows) - heterogeneous GROUP ALL /
  GROUP BY aggregate rows, same raw-SQL named-function treatment.

CONTEXT.md: add the "Graph Access Toolkit" Language entry after "Storage
Backend".
